Moving methods to the client class
